feat(DYLM-6): move coupled logic out of hook
